requirements.txt
torch
torchvision
ftfy
regex
tqdm
streamlit
matplotlib
scikit-learn
git+https://github.com/openai/CLIP.git
matplotlib
streamlit
numpy
pandas
Pillow
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
